SMSTracker/requirements.txt
Django==1.8.8
django-filter==0.13.0
django-geojson==2.9.0
django-guardian==1.4.1
django-leaflet==0.18.1
django-model-utils==2.5
django-notifications-hq==1.2
django-tables2==1.2.3
django-userena==2.0.1
djangorestframework==3.3.3
djangorestframework-jwt==1.8.0
easy-thumbnails==2.3
googlemaps==2.4.4
html2text==2014.12.29
jsonfield==1.0.3
Markdown==2.6.6
mysqlclient==1.3.7
Pillow==2.7.0
PyJWT==1.4.1
PyMySQL==0.7.6
pytz==2016.6
requests==2.10.0
six==1.10.0
yandex-maps==0.7
